encapsulate visitor logic
